[Feat] 컴포넌트 개발 및 footer.js삭제
- accrodion.html, accrodion.css, accrodion.js 아코디언 추가
- textlist.html 순차, 비순차 목록 컴포넌트 추가
- star.html 평점 컴포넌트 추가
- base.css 기본 속성 제거 추가
