+++ src/common/index.css
@@ -5,3 +5,4 @@
 @import '../components/style/footer.css';
 @import '../components/style/buttons.css';
 @import '../components/style/header.css';
+@import '../components/style/accrodion.css';
