+++ src/components/accrodion.html
@@ -1,0 +1,27 @@
+<!DOCTYPE html>
+<html lang="ko-KR">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>나이키 아코디언 컴포넌트</title>
+    <link rel="shortcut icon" href="/favicon/favicon.ico" type="image/x-icon" />
+    <link rel="icon" href="/favicon/favicon.svg" type="image/svg" />
+    <meta name="description" content="나이키 프로젝트 실습용 아코디언 컴포넌트 페이지입니다." />
+    <meta name="author" content="UI Project Team3" />
+    <link rel="preload" as="font" href="/font/woff2/PretendardVariable.woff2" crossorigin="anonymous" />
+    <link rel="stylesheet" as="style" href="/font/pretendardvariable.css" />
+    <link rel="stylesheet" href="../common/index.css" />
+    <link rel="stylesheet" href="./style/accrodion.css" />
+    <script type="module" src="../components/js/layout.js"></script>
+  </head>
+  <body>
+    <header id="header"></header>
+
+    <!-- 아코디언 컴포넌트 -->
+    <button type="button" class="btn-accrodion btn-arrow" aria-expanded="false">아코디언 제목</button>
+    <div class="detail-accrodion">아코디언 내용</div>
+    <!-- // 아코디언 컴포넌트 -->
+
+    <footer id="footer"></footer>
+  </body>
+</html>
