+++ src/components/footer.html
@@ -1,18 +1,18 @@
 <div class="footer-wrapper">
   <ul class="footer-menu">
     <li>
-      <button type="button" class="btn-arrow mo-block" aria-expanded="false"><span class="sr-only">나이키</span>안내</button>
+      <button type="button" class="btn-accrodion btn-arrow mo-block" aria-expanded="false"><span class="sr-only">나이키</span>안내</button>
       <span class="pc-block">안내</span>
-      <ul class="footer-sub-menu">
+      <ul class="detail-accrodion footer-sub-menu">
         <li><a href="/">멤버가입</a></li>
         <li><a href="/">매장찾기</a></li>
         <li><a href="/">나이키 저널</a></li>
       </ul>
     </li>
     <li>
-      <button type="button" class="btn-arrow mo-block" aria-expanded="false"><span class="sr-only">나이키</span>고객센터</button>
+      <button type="button" class="btn-accrodion btn-arrow mo-block" aria-expanded="false"><span class="sr-only">나이키</span>고객센터</button>
       <span class="pc-block">고객센터</span>
-      <ul class="footer-sub-menu">
+      <ul class="detail-accrodion footer-sub-menu">
         <li><a href="/">주문배송조회</a></li>
         <li><a href="/">반품 정책</a></li>
         <li><a href="/">결제 방법</a></li>
@@ -21,9 +21,9 @@
       </ul>
     </li>
     <li>
-      <button type="button" class="btn-arrow mo-block" aria-expanded="false"><span class="sr-only">나이키</span>회사소개</button>
+      <button type="button" class="btn-accrodion btn-arrow mo-block" aria-expanded="false"><span class="sr-only">나이키</span>회사소개</button>
       <span class="pc-block">회사소개</span>
-      <ul class="footer-sub-menu">
+      <ul class="detail-accrodion footer-sub-menu">
         <li><a href="/">About Nike</a></li>
         <li><a href="/">소식</a></li>
         <li><a href="/">채용</a></li>
