+++ src/components/js/layout.js
@@ -1,5 +1,5 @@
 import { headerFunction } from './header.js';
-import { footerFunction } from './footer.js';
+import { accrodionFunction } from './accrodion.js';
 
 const loadHTML = async (selector, url, callback) => {
   await fetch(url)
@@ -11,4 +11,4 @@ const loadHTML = async (selector, url, callback) => {
 };
 
 loadHTML('#header', '../components/header.html', headerFunction);
-loadHTML('#footer', '../components/footer.html', footerFunction);
+loadHTML('#footer', '../components/footer.html', accrodionFunction);
