+++ src/components/star.html
@@ -1,0 +1,33 @@
+<!DOCTYPE html>
+<html lang="ko-KR">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>나이키 평점 컴포넌트</title>
+    <link rel="shortcut icon" href="/favicon/favicon.ico" type="image/x-icon" />
+    <link rel="icon" href="/favicon/favicon.svg" type="image/svg" />
+    <meta name="description" content="나이키 프로젝트 실습용 평점 컴포넌트 페이지입니다." />
+    <meta name="author" content="UI Project Team3" />
+    <link rel="preload" as="font" href="/font/woff2/PretendardVariable.woff2" crossorigin="anonymous" />
+    <link rel="stylesheet" as="style" href="/font/pretendardvariable.css" />
+    <link rel="stylesheet" href="../common/index.css" />
+    <link rel="stylesheet" href="./style/star.css" />
+    <script type="module" src="../components/js/layout.js"></script>
+  </head>
+  <body>
+    <header id="header"></header>
+
+    <!-- 평점 컴포넌트 -->
+    <div class="star-rating" aria-hidden="true">
+      <span class="star fill"></span>
+      <span class="star fill"></span>
+      <span class="star fill"></span>
+      <span class="star"></span>
+      <span class="star"></span>
+    </div>
+    <span class="sr-only">만점 5점에 평점 3점</span>
+    <!-- // 평점 컴포넌트 -->
+
+    <footer id="footer"></footer>
+  </body>
+</html>
