+++ src/components/style/textlist.css
@@ -1,0 +1,18 @@
+/* 순차, 비순차 목록 컴포넌트 */
+.dot-list {
+  li {
+    position: relative;
+    padding-left: 1.6875rem;
+
+    &::before {
+      content: '';
+      position: absolute;
+      top: 0.8125rem;
+      left: 0.625rem;
+      width: 0.375rem;
+      height: 0.375rem;
+      border-radius: 50%;
+      background-color: var(--black);
+    }
+  }
+}
