+++ src/page/main.html
@@ -20,6 +20,9 @@
       <a href="../../index.html">메인 페이지 이동</a>
       <a href="./itemlist.html">아이템리스트 페이지 이동</a>
       <a href="./wishList.html">위시리스트 페이지 이동</a>
+      <a href="../components/accrodion.html">아코디언 컴포넌트 페이지 이동</a>
+      <a href="../components/textlist.html">순차, 비순차 목록 컴포넌트 페이지 이동</a>
+      <a href="../components/star.html">평점 컴포넌트 페이지 이동</a>
     </div>
     <br />
     <h1></h1>
